+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/BlameLineDto.java
@@ -21,28 +21,21 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import React from "react";
-import DateFromNow from "./DateFromNow";
-import { storiesOf } from "@storybook/react";
 
-const baseProps = {
-  timeZone: "Europe/Berlin",
-  baseDate: "2019-10-12T13:56:42+02:00"
-};
+package sonia.scm.api.v2.resources;
 
-storiesOf("DateFromNow", module).add("Default", () => (
-  <div>
-    <p>
-      <DateFromNow date="2009-06-30T18:30:00+02:00" {...baseProps} />
-    </p>
-    <p>
-      <DateFromNow date="2019-06-30T18:30:00+02:00" {...baseProps} />
-    </p>
-    <p>
-      <DateFromNow date="2019-10-12T13:56:40+02:00" {...baseProps} />
-    </p>
-    <p>
-      <DateFromNow date="2019-10-11T13:56:40+02:00" {...baseProps} />
-    </p>
-  </div>
-));
+import lombok.Data;
+
+import java.time.Instant;
+
+@Data
+public class BlameLineDto {
+
+  private PersonDto author;
+  private String code;
+  private String description;
+  private int lineNumber;
+  private String revision;
+  private Instant when;
+
+}